none block with rest api

+++ microservices/recommendation-service/src/main/java/com/example/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 @Slf4j
 @RestController
@@ -22,35 +22,37 @@ public class RecommendationServiceImpl implements RecommendationService {
     private final RecommendationMapper mapper;
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
-        try {
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity newEntity = repository.save(entity);
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
+        if (body.getProductId() < 1) throw new InvalidInputException("Invalid productId: " + body.getProductId());
 
-            log.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-            return mapper.entityToApi(newEntity);
+        RecommendationEntity entity = mapper.apiToEntity(body);
 
-        } catch (DuplicateKeyException dke) {
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId());
-        }
+        return repository.save(entity)
+                .log()
+                .onErrorMap(
+                        DuplicateKeyException.class,
+                        ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId()))
+                .map(mapper::entityToApi);
     }
 
     @Override
-    public void deleteRecommendation(int productId) {
-        log.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
-        repository.deleteAll(repository.findByProductId(productId));
-    }
-
-    @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = mapper.entityListToApiList(entityList);
-        list.forEach(it -> it.setServiceAddress(serviceUtil.getServiceAddress()));
+        return repository.findByProductId(productId)
+                .log()
+                .map(mapper::entityToApi)
+                .map(it -> {
+                    it.setServiceAddress(serviceUtil.getServiceAddress());
+                    return it;
+                });
+    }
 
-        log.debug("getRecommendations: response size: {}", list.size());
+    @Override
+    public Mono<Void> deleteRecommendation(int productId) {
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
-        return list;
+        log.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
+        return repository.deleteAll(repository.findByProductId(productId));
     }
 }
